refactor : util 패키지의 목적에 맞게 도메인 로직을 도메인 패키지로 이동한다
